app/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
python-dotenv==1.0.1
groq==0.11.0
google-generativeai==0.8.3
pymupdf==1.24.11
pdfplumber==0.11.4
alembic==1.13.3
